chore: get rid of a small bit of extra added additional redundancy
