+++ packages/kernel/test/storage.ts
@@ -1,5 +1,7 @@
 import type { KVStore, KernelDatabase, VatStore } from '@ocap/store';
 
+import { keySearch } from '../src/utils/key-search.ts';
+
 /* eslint-disable no-lonely-if, no-else-return */ // stupid rules that encourage unclear code
 
 /**
@@ -24,43 +26,6 @@ function makeMapKVStoreInternal(map: Map<string, string>): KVStore {
   let lastNextKey: string | null = null;
   let lastNextKeyIndex: number = -1;
 
-  /**
-   * Binary search for key position.
-   *
-   * @param key - The key to search `keyCache` for.
-   *
-   * @returns the index into `keyCache` for the first key that is greater than
-   *   `key`, or -1 if no such key exists.
-   */
-  function search(key: string): number {
-    if (keyCache === null) {
-      return -1;
-    }
-    let beg = 0;
-    let end = keyCache.length - 1;
-    if (key < (keyCache[beg] as string)) {
-      return beg;
-    }
-    if ((keyCache[end] as string) < key) {
-      return -1;
-    }
-    while (beg <= end) {
-      const mid = Math.floor((beg + end) / 2);
-      if (keyCache[mid] === key) {
-        return mid;
-      }
-      if (key < (keyCache[mid] as string)) {
-        end = mid - 1;
-      } else {
-        beg = mid + 1;
-      }
-      if (beg === end) {
-        return beg;
-      }
-    }
-    return -1;
-  }
-
   return {
     get(key: string): string | undefined {
       return map.get(key);
@@ -69,7 +34,8 @@ function makeMapKVStoreInternal(map: Map<string, string>): KVStore {
       if (keyCache === null) {
         keyCache = Array.from(map.keys()).sort();
       }
-      const index = lastNextKey === key ? lastNextKeyIndex : search(key);
+      const index =
+        lastNextKey === key ? lastNextKeyIndex : keySearch(keyCache, key);
       if (index < 0) {
         lastNextKey = null;
         lastNextKeyIndex = -1;
